BACKEND/requirements.txt
APScheduler==3.11.0
fastapi==0.115.12
pydantic==2.11.3
python-dotenv==1.1.0
python_jose==3.4.0
Requests==2.32.3
SQLAlchemy==2.0.40
